requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
amqp==5.3.1
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==24.2.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via
    #   -r requirements/base.txt
    #   celery
bytecode==0.16.0
    # via ddtrace
celery==5.4.0
    # via -r requirements/base.txt
certifi==2024.8.30
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.4.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.1
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==1.8.2
    # via
    #   -r requirements/base.txt
    #   -r requirements/test.in
    #   edx-toggles
coverage[toml]==7.6.7
    # via pytest-cov
cryptography==43.0.3
    # via
    #   -r requirements/base.txt
    #   pyjwt
ddt==1.7.2
    # via -r requirements/test.in
ddtrace==2.16.4
    # via -r requirements/test.in
deprecated==1.2.15
    # via opentelemetry-api
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-toggles
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
djangorestframework==3.15.2
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-codejail==3.5.2
    # via -r requirements/base.txt
edx-django-utils==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-toggles
edx-drf-extensions==10.5.0
    # via -r requirements/base.txt
edx-opaque-keys==2.11.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-toggles==5.2.0
    # via -r requirements/base.txt
envier==0.6.1
    # via ddtrace
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
importlib-metadata==8.5.0
    # via opentelemetry-api
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via
    #   -r requirements/base.txt
    #   code-annotations
jsonschema==4.23.0
    # via -r requirements/base.txt
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
kombu==5.4.2
    # via
    #   -r requirements/base.txt
    #   celery
markupsafe==3.0.2
    # via
    #   -r requirements/base.txt
    #   jinja2
newrelic==10.3.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
opentelemetry-api==1.28.2
    # via ddtrace
packaging==24.2
    # via pytest
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.48
    # via
    #   -r requirements/base.txt
    #   click-repl
protobuf==5.28.3
    # via ddtrace
psutil==6.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pyjwt[crypto]==2.10.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pymongo==4.10.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.3
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-randomly
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
pytest-randomly==3.16.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   celery
python-slugify==8.0.4
    # via
    #   -r requirements/base.txt
    #   code-annotations
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
referencing==0.35.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
rpds-py==0.21.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   edx-codejail
    #   python-dateutil
sqlparse==0.5.2
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via
    #   -r requirements/base.txt
    #   python-slugify
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   ddtrace
    #   edx-opaque-keys
tzdata==2024.2
    # via
    #   -r requirements/base.txt
    #   celery
    #   kombu
urllib3==2.2.3
    # via
    #   -r requirements/base.txt
    #   requests
vine==5.1.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
wrapt==1.16.0
    # via
    #   ddtrace
    #   deprecated
xmltodict==0.14.2
    # via ddtrace
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==75.6.0
    # via -r requirements/base.txt
